requirements.txt
altair==5.5.0
attrs==25.3.0
bcc==0.18.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
gitdb==4.0.12
GitPython==3.1.44
gyp==0.1
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
more-itertools==8.10.0
narwhals==1.41.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pop-transition==1.1.2
protobuf==6.31.1
psutil==5.9.0
pyarrow==20.0.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydbus==0.6.0
pydeck==0.9.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyte==0.8.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
rpds-py==0.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
smmap==5.0.2
streamlit==1.45.1
systemd-python==234
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
watchdog==6.0.0
wcwidth==0.2.5
xdg==5
xkit==0.0.0
zipp==1.0.0
